Struct picea::Context

pub struct Context<'a, S: 'a, P: 'a, N: 'a + Node<S, P> + ?Sized> { /* fields omitted */ }

Provides an interface to the tree structure.

Methods

impl<'a, S, P, N: Node<S, P>> Context<'a, S, P, N>

pub fn send(&mut self, event: S)

Send an event to the parent of this node.

pub fn send_all<I: Iterator<Item = S>>(&mut self, events: I)

Send several events to the parent of this node.

pub fn accept(&mut self, event: N::Event)

Send an event to self.

pub fn accept_all<I: Iterator<Item = N::Event>>(&mut self, events: I)

Send several events to self.

pub fn children(&mut self) -> TreeBuilder<N::Event, N::Output>

Build child nodes.

pub fn siblings(&mut self) -> TreeBuilder<S, P>

Build sibling nodes.

pub fn kill(&mut self)

Destroy this node once the current update cycle ends. All events and the end call for the current cycle are still made.

pub fn close<F>(&mut self, close: F) where
F: FnOnce(Box<N>, Tree<N::Event, N::Output>) + 'static,

Move this node once the current update cycle ends. All events and the end call for the current cycle are still made.

pub fn close_boxed(
&mut self,
close: Box<FnBox(Box<N>, Tree<N::Event, N::Output>)>
)

Move this node once the current update cycle ends. All events and the end call for the current cycle are still made.

pub fn revive(&mut self)

Cancels the destruction of this node.

pub fn param(&self) -> &P

Get an immutable reference to the input parameter specified by the parent.