Trait picea::Node

pub trait Node<S, P>: Sized {
    type Output;
    type Event;
    fn update(&mut self, ctx: &mut Context<S, P, Self>) -> Self::Output;

    fn event(&mut self, ctx: &mut Context<S, P, Self>, event: Self::Event);

    fn end(&mut self, ctx: &mut Context<S, P, Self>);
}

A structure that can be used as a node in a tree.

Associated Types

type Output

The parameter type that this node generates.

type Event

The event type that this node accepts.

Required Methods

fn update(&mut self, ctx: &mut Context<S, P, Self>) -> Self::Output

The update cycle for a node starts with this function being called. The cycles for all children will be completed immediately after this function exits.

fn event(&mut self, ctx: &mut Context<S, P, Self>, event: Self::Event)

This function might be called several times after update and before end.

fn end(&mut self, ctx: &mut Context<S, P, Self>)

The update cycle for a node ends with this function being called.